Moved example tests into Example/ folder
