+++ tests/Example/PHPUnitExampleTest.php
@@ -1,10 +1,10 @@
 <?php
 /**
- * Example Test
+ * PHPUnit Example Test
  * 
  * @author Leonid Mamchenkov <l.mamchenkov@qobo.biz>
  */
-class ExampleTest extends PHPUnit_Framework_TestCase {
+class PHPUnitExampleTest extends PHPUnit_Framework_TestCase {
 
 	/**
 	 * Example unit test
